compute: refactoring the validate functions into their own package
