+++ azurerm/internal/services/compute/ssh_public_key_resource.go
@@ -6,6 +6,8 @@ import (
 	"regexp"
 	"time"
 
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/compute/validate"
+
 	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2020-12-01/compute"
 	"github.com/hashicorp/go-azure-helpers/response"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
@@ -60,7 +62,7 @@ func resourceSshPublicKey() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     false,
-				ValidateFunc: ValidateSSHKey,
+				ValidateFunc: validate.SSHKey,
 			},
 
 			"tags": tags.Schema(),
