+++ azurerm/internal/services/compute/validate/dedicated_host_group_name.go
@@ -1,0 +1,11 @@
+package validate
+
+import (
+	"regexp"
+
+	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
+)
+
+func DedicatedHostGroupName() func(i interface{}, k string) (warnings []string, errors []error) {
+	return validation.StringMatch(regexp.MustCompile(`^[^_\W][\w-.]{0,78}[\w]$`), "")
+}
