+++ azurerm/internal/services/compute/validate/dedicated_host_name.go
@@ -1,0 +1,5 @@
+package validate
+
+func DedicatedHostName() func(i interface{}, k string) (warnings []string, errors []error) {
+	return DedicatedHostGroupName()
+}
