+++ azurerm/internal/services/compute/validate/disk_size_gb.go
@@ -1,0 +1,12 @@
+package validate
+
+import "fmt"
+
+func DiskSizeGB(v interface{}, _ string) (warnings []string, errors []error) {
+	value := v.(int)
+	if value < 0 || value > 32767 {
+		errors = append(errors, fmt.Errorf(
+			"The `disk_size_gb` can only be between 0 and 32767"))
+	}
+	return warnings, errors
+}
