+++ azurerm/internal/services/compute/virtual_machine_scale_set_resource.go
@@ -7,6 +7,8 @@ import (
 	"strings"
 	"time"
 
+	validate2 "github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/compute/validate"
+
 	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2020-12-01/compute"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/structure"
@@ -631,7 +633,7 @@ func resourceVirtualMachineScaleSet() *schema.Resource {
 							Type:         schema.TypeInt,
 							Optional:     true,
 							Computed:     true,
-							ValidateFunc: validateDiskSizeGB,
+							ValidateFunc: validate2.DiskSizeGB,
 						},
 
 						"managed_disk_type": {
